requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
Django==4.2.3
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-redis==5.3.0
django-widget-tweaks==1.4.12
idna==3.4
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
